requirements.txt
agno==1.1.4
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiohttp-client-cache==0.11.1
aiosignal==1.3.2
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.8.0
async-lru==2.0.4
async-timeout==5.0.1
attrs==25.1.0
babel==2.17.0
backoff==2.2.1
backports.tarfile==1.2.0
beautifulsoup4==4.13.3
Brotli==1.1.0
build==1.2.2.post1
CacheControl==0.14.2
cattrs==24.1.2
certifi==2025.1.31
charset-normalizer==3.4.1
cleo==2.1.0
click==8.1.8
colorama==0.4.6
courlan==1.3.2
crashtest==0.4.1
dateparser==1.2.1
deepdiff==8.2.0
defusedxml==0.8.0rc2
distlib==0.3.9
distro==1.9.0
docstring_parser==0.16
duckduckgo_search==7.4.2
dulwich==0.22.7
et_xmlfile==2.0.0
exceptiongroup==1.2.2
fastapi==0.115.8
fastjsonschema==2.21.1
filelock==3.17.0
findpython==0.6.2
frozendict==2.4.6
frozenlist==1.5.0
gitdb==4.0.12
GitPython==3.1.44
groq==0.18.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
html5lib==1.1
htmldate==1.9.3
httpcore==1.0.7
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.6.1
inscriptis==2.5.3
installer==0.7.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jiter==0.8.2
jusText==3.0.1
keyring==25.6.0
lxml==5.3.1
lxml_html_clean==0.4.1
markdown-it-py==3.0.0
mdurl==0.1.2
monotonic==1.6
more-itertools==10.6.0
msgpack==1.1.0
multidict==6.1.0
multitasking==0.0.11
numpy==2.0.2
openbb==4.4.0
openbb-benzinga==1.4.0
openbb-bls==1.1.0
openbb-cftc==1.1.0
openbb-commodity==1.3.0
openbb-core==1.4.0
openbb-crypto==1.4.0
openbb-currency==1.4.0
openbb-derivatives==1.4.0
openbb-econdb==1.3.0
openbb-economy==1.4.0
openbb-equity==1.4.0
openbb-etf==1.4.0
openbb-federal-reserve==1.4.0
openbb-fixedincome==1.4.1
openbb-fmp==1.4.0
openbb-fred==1.4.0
openbb-imf==1.1.0
openbb-index==1.4.0
openbb-intrinio==1.4.0
openbb-news==1.4.0
openbb-oecd==1.4.0
openbb-platform-api==1.1.2
openbb-polygon==1.4.0
openbb-regulators==1.4.0
openbb-sec==1.4.0
openbb-tiingo==1.4.0
openbb-tradingeconomics==1.4.0
openbb-us-eia==1.1.0
openbb-yfinance==1.4.0
openpyxl==3.1.5
orderly-set==5.3.0
packaging==24.2
pandas==2.2.3
pbs-installer==2025.2.12
peewee==3.17.9
persian-names==1.5.1
phidata==2.7.10
pkginfo==1.12.0
platformdirs==4.3.6
poetry==2.1.1
poetry-core==2.1.1
posthog==3.13.0
propcache==0.2.1
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.10.1
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.18
pytz==2025.1
pywin32-ctypes==0.2.3
PyYAML==6.0.2
RapidFuzz==3.12.1
regex==2024.11.6
requests==2.32.3
requests-cache==1.2.1
requests-toolbelt==1.0.0
rich==13.9.4
ruff==0.7.4
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.6
starlette==0.45.3
tld==0.13
tomli==2.2.1
tomlkit==0.13.2
tqdm==4.67.1
trafilatura==2.0.0
trove-classifiers==2025.1.15.22
typer==0.15.1
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3
url-normalize==1.4.3
urllib3==2.3.0
uuid7==0.1.0
uvicorn==0.34.0
virtualenv==20.29.2
webencodings==0.5.1
websockets==14.2
xlrd==2.0.1
xmltodict==0.13.0
yarl==1.18.3
yfinance==0.2.53
zipp==3.21.0
zstandard==0.23.0
